+++ ethereum/evmtool/src/main/java/org/hyperledger/besu/evmtool/B11rSubCommand.java
@@ -32,6 +32,9 @@ import java.io.InputStreamReader;
 import java.io.PrintStream;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Stack;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -41,7 +44,12 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.apache.tuweni.bytes.Bytes;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
+import picocli.CommandLine.IParameterConsumer;
+import picocli.CommandLine.Model.ArgSpec;
+import picocli.CommandLine.Model.CommandSpec;
 import picocli.CommandLine.Option;
+import picocli.CommandLine.Parameters;
+import picocli.CommandLine.ParentCommand;
 
 @Command(
     name = COMMAND_NAME,
@@ -74,6 +82,12 @@ public class B11rSubCommand implements Runnable {
   private final Path ommers = stdinPath;
 
   @Option(
+      names = {"--input.withdrawals"},
+      paramLabel = "full path",
+      description = "The withdrawals for the block")
+  private final Path withdrawals = stdinPath;
+
+  @Option(
       names = {"--seal.clique"},
       paramLabel = "full path",
       description = "The clique seal/signature for the block")
@@ -104,7 +118,25 @@ public class B11rSubCommand implements Runnable {
       description = "The account state after the transition")
   private final Path outBlock = Path.of("block.json");
 
-  @CommandLine.ParentCommand private final EvmToolCommand parentCommand;
+  @ParentCommand private final EvmToolCommand parentCommand;
+
+  @Parameters(parameterConsumer = OnlyEmptyParams.class)
+  @SuppressWarnings("UnusedVariable")
+  private final List<String> parameters = new ArrayList<>();
+
+  static class OnlyEmptyParams implements IParameterConsumer {
+    @Override
+    public void consumeParameters(
+        final Stack<String> args, final ArgSpec argSpec, final CommandSpec commandSpec) {
+      while (!args.isEmpty()) {
+        if (!args.pop().isEmpty()) {
+          throw new CommandLine.ParameterException(
+              argSpec.command().commandLine(),
+              "The block-builder command does not accept any non-empty parameters");
+        }
+      }
+    }
+  }
 
   @SuppressWarnings("unused")
   public B11rSubCommand() {
@@ -121,39 +153,46 @@ public class B11rSubCommand implements Runnable {
   @Override
   public void run() {
     ObjectMapper objectMapper = JsonUtils.createObjectMapper();
-    final ObjectReader t8nReader = objectMapper.reader();
+    final ObjectReader b11rReader = objectMapper.reader();
 
     ObjectNode config;
     try {
       if (header.equals(stdinPath)
           || txs.equals(stdinPath)
           || ommers.equals(stdinPath)
-          || sealClique.equals(stdinPath)) {
+          || sealClique.equals(stdinPath)
+          || withdrawals.equals(stdinPath)) {
         config =
             (ObjectNode)
-                t8nReader.readTree(new InputStreamReader(parentCommand.in, StandardCharsets.UTF_8));
+                b11rReader.readTree(
+                    new InputStreamReader(parentCommand.in, StandardCharsets.UTF_8));
       } else {
         config = objectMapper.createObjectNode();
       }
 
       if (!header.equals(stdinPath)) {
         try (FileReader reader = new FileReader(header.toFile(), StandardCharsets.UTF_8)) {
-          config.set("header", t8nReader.readTree(reader));
+          config.set("header", b11rReader.readTree(reader));
         }
       }
       if (!txs.equals(stdinPath)) {
         try (FileReader reader = new FileReader(txs.toFile(), StandardCharsets.UTF_8)) {
-          config.set("txs", t8nReader.readTree(reader));
+          config.set("txs", b11rReader.readTree(reader));
+        }
+      }
+      if (!withdrawals.equals(stdinPath)) {
+        try (FileReader reader = new FileReader(withdrawals.toFile(), StandardCharsets.UTF_8)) {
+          config.set("withdrawals", b11rReader.readTree(reader));
         }
       }
       if (!ommers.equals(stdinPath)) {
         try (FileReader reader = new FileReader(ommers.toFile(), StandardCharsets.UTF_8)) {
-          config.set("ommers", t8nReader.readTree(reader));
+          config.set("ommers", b11rReader.readTree(reader));
         }
       }
       if (!sealClique.equals(stdinPath)) {
         try (FileReader reader = new FileReader(sealClique.toFile(), StandardCharsets.UTF_8)) {
-          config.set("clique", t8nReader.readTree(reader));
+          config.set("clique", b11rReader.readTree(reader));
         }
       }
     } catch (final JsonProcessingException jpe) {