+++ ethereum/evmtool/src/main/java/org/hyperledger/besu/evmtool/T8nSubCommand.java
@@ -30,6 +30,7 @@ import org.hyperledger.besu.ethereum.core.Transaction;
 import org.hyperledger.besu.ethereum.referencetests.ReferenceTestEnv;
 import org.hyperledger.besu.ethereum.referencetests.ReferenceTestWorldState;
 import org.hyperledger.besu.ethereum.rlp.BytesValueRLPInput;
+import org.hyperledger.besu.evm.AccessListEntry;
 import org.hyperledger.besu.evm.tracing.OperationTracer;
 import org.hyperledger.besu.evm.tracing.StandardJsonTracer;
 import org.hyperledger.besu.plugin.data.TransactionType;
@@ -48,6 +49,10 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.Spliterator;
+import java.util.Spliterators;
+import java.util.Stack;
+import java.util.stream.StreamSupport;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -58,7 +63,12 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import picocli.CommandLine.Command;
+import picocli.CommandLine.IParameterConsumer;
+import picocli.CommandLine.Model.ArgSpec;
+import picocli.CommandLine.Model.CommandSpec;
 import picocli.CommandLine.Option;
+import picocli.CommandLine.ParameterException;
+import picocli.CommandLine.Parameters;
 import picocli.CommandLine.ParentCommand;
 
 @Command(
@@ -139,6 +149,24 @@ public class T8nSubCommand implements Runnable {
   private String rewardString = null;
 
   @ParentCommand private final EvmToolCommand parentCommand;
+
+  @Parameters(parameterConsumer = OnlyEmptyParams.class)
+  @SuppressWarnings("UnusedVariable")
+  private final List<String> parameters = new ArrayList<>();
+
+  static class OnlyEmptyParams implements IParameterConsumer {
+    @Override
+    public void consumeParameters(
+        final Stack<String> args, final ArgSpec argSpec, final CommandSpec commandSpec) {
+      while (!args.isEmpty()) {
+        if (!args.pop().isEmpty()) {
+          throw new ParameterException(
+              argSpec.command().commandLine(),
+              "The transition command does not accept any non-empty parameters");
+        }
+      }
+    }
+  }
 
   @SuppressWarnings("unused")
   public T8nSubCommand() {
@@ -352,6 +380,29 @@ public class T8nSubCommand implements Runnable {
                 new BigInteger(
                     1,
                     Bytes.fromHexStringLenient(txNode.get("chainId").textValue()).toArrayUnsafe()));
+          }
+
+          if (txNode.has("accessList")) {
+            JsonNode accessList = txNode.get("accessList");
+            if (!accessList.isArray()) {
+              parentCommand.out.printf(
+                  "TX json node unparseable: expected accessList to be an array - %s%n", txNode);
+              continue;
+            }
+            List<AccessListEntry> entries = new ArrayList<>(accessList.size());
+            for (JsonNode entryAsJson : accessList) {
+              Address address = Address.fromHexString(entryAsJson.get("address").textValue());
+              List<String> storageKeys =
+                  StreamSupport.stream(
+                          Spliterators.spliteratorUnknownSize(
+                              entryAsJson.get("storageKeys").elements(), Spliterator.ORDERED),
+                          false)
+                      .map(JsonNode::textValue)
+                      .toList();
+              var accessListEntry = AccessListEntry.createAccessListEntry(address, storageKeys);
+              entries.add(accessListEntry);
+            }
+            builder.accessList(entries);
           }
 
           if (txNode.has("secretKey")) {